update build function idioms
