+++ src/layers/learning/CMakeLists.txt
@@ -1,10 +1,12 @@
 # Add the source files for this directory
 set_full_path(THIS_DIR_SOURCES
   channelwise_scale_bias.cpp
+  channelwise_scale_bias_builder.cpp
   convolution.cpp
   deconvolution.cpp
   entrywise_scale_bias.cpp
   embedding.cpp
+  embedding_builder.cpp
   fully_connected.cpp
   )
