+++ src/layers/learning/embedding.cpp
@@ -27,9 +27,6 @@
 #define LBANN_EMBEDDING_LAYER_INSTANTIATE
 #include "lbann/layers/learning/embedding.hpp"
 
-#include <lbann/proto/proto_common.hpp>
-#include <lbann.pb.h>
-
 namespace lbann {
 
 template <typename TensorDataType, data_layout Layout, El::Device Device>
@@ -109,46 +106,6 @@ void embedding_layer<TensorDataType, Layout, Device>::bp_compute() {
 
 }
 
-namespace
-{
-template <typename T, data_layout L, El::Device D>
-struct Builder
-{
-  static std::unique_ptr<Layer> Build(...)
-  {
-    LBANN_ERROR("Attempted to instantiate layer \"embedding\""
-                "with Layout=", to_string(L), ".\nThis layer is only "
-                "supported with DATA_PARALLEL data layout.");
-    return nullptr;
-  }
-};
-
-template <typename TensorDataType, El::Device Device>
-struct Builder<TensorDataType,data_layout::DATA_PARALLEL,Device>
-{
-  template <typename... Args>
-  static std::unique_ptr<Layer> Build(Args&&... args)
-  {
-    using LayerType = embedding_layer<TensorDataType,data_layout::DATA_PARALLEL,Device>;
-    return make_unique<LayerType>(std::forward<Args>(args)...);
-  }
-};
-} // namespace <anon>
-
-template <typename TensorDataType, data_layout Layout, El::Device Device>
-std::unique_ptr<Layer> build_embedding_layer_from_pbuf(
-  lbann_comm* comm, lbann_data::Layer const& proto_layer)
-{
-  using BuilderType = Builder<TensorDataType, Layout, Device>;
-  LBANN_ASSERT_MSG_HAS_FIELD(proto_layer, embedding);
-
-  const auto& params = proto_layer.embedding();
-  const size_t num_embeddings = params.num_embeddings();
-  const size_t embedding_dim = params.embedding_dim();
-  const El::Int padding_idx = (params.has_padding_idx() ?
-                               params.padding_idx().value() : -1);
-  return BuilderType::Build(comm, num_embeddings, embedding_dim, padding_idx);
-}
 
 // Explicit instantiation
 #define PROTO(T)                                                        \
@@ -156,12 +113,5 @@ std::unique_ptr<Layer> build_embedding_layer_from_pbuf(
 
 #define LBANN_INSTANTIATE_CPU_HALF
 #include "lbann/macros/instantiate.hpp"
-#undef PROTO
-
-#define PROTO_DEVICE(T, Device) \
-  LBANN_LAYER_BUILDER_ETI(embedding, T, Device)
-
-#define LBANN_INSTANTIATE_GPU_HALF
-#include "lbann/macros/instantiate_device.hpp"
 
 } // namespace lbann
