+++ src/layers/transform/CMakeLists.txt
@@ -3,8 +3,10 @@ set_full_path(THIS_DIR_SOURCES
   bernoulli.cpp
   categorical_random.cpp
   concatenate.cpp
+  concatenate_builder.cpp
   constant.cpp
   crop.cpp
+  crop_builder.cpp
   discrete_random.cpp
   dummy.cpp
   evaluation.cpp
